Removed unneeded dependencies
