+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
     "typescript.tsdk": "node_modules/typescript/lib",
-    "editor.formatOnSave": true
+    "editor.formatOnSave": true,
+    "eslint.enable": false
 }
