+++ package.json
@@ -25,11 +25,8 @@
     },
     "homepage": "https://skyaaron.com/",
     "dependencies": {
-        "@mdx-js/mdx": "^1.0.21",
-        "@mdx-js/react": "^1.0.21",
         "gatsby": "^2.0.0",
         "gatsby-plugin-favicon": "^3.1.0",
-        "gatsby-plugin-mdx": "^1.0.6",
         "gatsby-plugin-nprogress": "^2.0.0",
         "gatsby-plugin-react-helmet": "^3.0.5",
         "gatsby-plugin-remove-trailing-slashes": "^2.0.0",
@@ -51,6 +48,7 @@
         "@types/react": "^16.7.0",
         "@types/react-dom": "^16.0.0",
         "@types/react-helmet": "^5.0.3",
+        "@types/html-pdf": "^2.1.2",
         "codecov": "^3.0.0",
         "graphql-code-generator": "^0.15.0",
         "html-pdf": "^2.2.0",
